requirements.txt
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
ccimport==0.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cumm-cu118==0.7.11
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
fastprogress==1.0.3
fire==0.7.0
Flask==3.0.3
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
laspy==2.5.4
lazrs==0.6.2
llvmlite==0.43.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
memory-profiler==0.61.0
mistune==3.1.1
munch==4.0.0
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.3
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
open3d==0.19.0
overrides==7.7.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.16
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.1
pyparsing==3.2.1
pyransac3d==0.6.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
pywin32==308
pywinpty==2.0.15
pyyaml
pyzmq==26.2.1
referencing==0.36.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.3
sniffio==1.3.1
soupsieve==2.6
spconv-cu118==2.3.8
stack-data==0.6.3
sympy==1.13.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.14
tinycss2==1.4.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
uri-template==1.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zstandard==0.23.0
